+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Modified log messages about ONTAP media types not associated with performance classes (Issue [#158](https://github.com/NetApp/trident/issues/158)).
 
 **Enhancements:**
+- Added ability to set snapshotReserve in backend config files, volume creation options, and PVC annotations (Issue [#43](https://github.com/NetApp/trident/issues/43)).
 
 ## v18.07.0